Модернизация интерфейса INotifications. Увеличено максимальное количество типов уведомлений с 8 до 16. Добавлены новые типы уведомлений - в окне со вкладками и в панели задач.

+++ trunk/src/plugins/notifications/notifykindswidget.h
@@ -7,13 +7,13 @@
 #include "ui_notifykindswidget.h"
 
 class NotifyKindsWidget :
-			public QWidget,
-			public IOptionsWidget
+	public QWidget,
+	public IOptionsWidget
 {
 	Q_OBJECT;
 	Q_INTERFACES(IOptionsWidget);
 public:
-	NotifyKindsWidget(INotifications *ANotifications, const QString &AId, const QString &ATitle, uchar AKindMask, QWidget *AParent);
+	NotifyKindsWidget(INotifications *ANotifications, const QString &AId, const QString &ATitle, ushort AKindMask, QWidget *AParent);
 	~NotifyKindsWidget();
 	virtual QWidget* instance() { return this; }
 public slots:
